Add Eq, Ord, Num, Enum, and Bounded instances for SomeSing
