+++ singletons.cabal
@@ -97,6 +97,7 @@ library
                       Data.Singletons.Deriving.Enum,
                       Data.Singletons.Deriving.Ord,
                       Data.Singletons.Deriving.Show,
+                      Data.Singletons.Internal,
                       Data.Singletons.Prelude.List.NonEmpty.Internal,
                       Data.Singletons.Promote,
                       Data.Singletons.Promote.Monad,
